feat(crates-io): trim renaming deps
